SMS/requirements.txt
Django>=4.0
django-crispy-forms
crispy-bootstrap5
python-dotenv
weasyprint>=60.0
